Refactor DB entities to use long for numeric fields
Changed various entity properties from int to long for better compatibility with large values, including Clipping, Vocab, Lookup, MetaData, Version, and Word. Updated repository and service logic to use long types and added GetSafeLong helper for database reads. Also refactored repository interfaces to return bool for Add/Update/Delete methods and improved error handling and null safety in repository implementations.

+++ KindleMate2.Application/Services/KM2DB/KMDatabaseService.cs
@@ -165,9 +165,9 @@ namespace KindleMate2.Application.Services.KM2DB {
                 } else if (metadata.Contains("文章剪切") || metadata.Contains("Cut")) {
                     brieftype = BriefType.Cut;
                 }
-                clipping.BriefType = brieftype;
+                clipping.BriefType = (long)brieftype;
 
-                if (clipping.BriefType == BriefType.Bookmark) {
+                if (clipping.BriefType == (long)BriefType.Bookmark) {
                     continue;
                 }
 
@@ -301,7 +301,7 @@ namespace KindleMate2.Application.Services.KM2DB {
                 }
                 _clippingRepository.UpdateBriefTypeByKey(new Clipping {
                     Key = clipping.Key,
-                    BriefType = BriefType.Hide
+                    BriefType = (long)BriefType.Hide
                 });
                 return true;
             } catch (Exception e) {
