+++ KindleMate2.Domain/Entities/VocabDB/Lookup.cs
@@ -1,6 +1,6 @@
 ﻿namespace KindleMate2.Domain.Entities.VocabDB {
     public class Lookup {
-        private int _timestamp;
+        private long _timestamp;
         public required string Id { get; set; } = null!;
         public string? WordKey { get; set; }
         public string? BookKey { get; set; }
@@ -8,7 +8,7 @@
         public string? Pos { get; set; }
         public string? Usage { get; set; }
 
-        public int? Timestamp {
+        public long? Timestamp {
             get => _timestamp;
             set => _timestamp = value ?? 0;
         }
