+++ KindleMate2.Domain/Entities/VocabDB/MetaData.cs
@@ -2,7 +2,7 @@
     public class MetaData {
         public required string Id { get; set; } = null!;
         public string? Dsname { get; set; }
-        public int? Sscnt { get; set; }
+        public long? Sscnt { get; set; }
         public string? Profileid { get; set; }
     }
 }
